refactor: address PR review feedback
- Consolidate 3 kubeconfig proxy routes into single [...path] catch-all
- Rename MergedCluster type to ConnectedCluster for clarity
- Rename kubectl_clusters_tool.py to list_clusters_tool.py
- Rename _make_vault_provider to _make_vault_key
- Remove atexit.register (redundant with finally block, memory leak)
- Use shlex.split() instead of str.split() for proper quoted arg handling
