--- client/src/app/api/kubeconfig/clusters/route.ts
@@ -1,6 +1,0 @@
-import { NextRequest } from 'next/server'
-import { forwardRequest } from '@/lib/backend-proxy'
-
-export async function GET(request: NextRequest) {
-  return forwardRequest(request, 'GET', '/api/kubeconfig/clusters', 'kubeconfig clusters')
-}
