--- client/src/app/api/kubeconfig/upload/route.ts
@@ -1,6 +1,0 @@
-import { NextRequest } from 'next/server'
-import { forwardRequest } from '@/lib/backend-proxy'
-
-export async function POST(request: NextRequest) {
-  return forwardRequest(request, 'POST', '/api/kubeconfig/upload', 'kubeconfig upload')
-}
