+++ server/chat/backend/agent/tools/list_clusters_tool.py
